Changed to use separate AppNavigators for work with Activities. Changed some javadocs. Updated some lib versions.
