+++ library/src/main/java/ru/terrakok/cicerone/Router.java
@@ -121,8 +121,8 @@ public class Router extends BaseRouter {
     }
 
     /**
-     * Full completion the chain and exit.
-     * It's usually necessary to complete the application or embedded navigation.
+     * Remove all screens from the chain and exit.
+     * It's mostly used to finish the application or close a supplementary navigation chain.
      */
     public void finishChain() {
         executeCommand(new BackTo(null));
