+++ sample/build.gradle
@@ -6,7 +6,7 @@ repositories {
 
 android {
     compileSdkVersion 25
-    buildToolsVersion '25'
+    buildToolsVersion '25.0.2'
 
     defaultConfig {
         minSdkVersion 16
@@ -29,13 +29,14 @@ android {
 
 ext {
     supportLibraryVersion = "25.0.1"
-    moxyVersion = "1.2.1"
-    daggerVersion = "2.8"
+    moxyVersion = "1.4.6"
+    daggerVersion = "2.9"
 }
 
 dependencies {
     // Support libraries
     compile "com.android.support:appcompat-v7:$supportLibraryVersion"
+    compile "com.android.support:design:$supportLibraryVersion"
 
     //MVP Moxy
     compile "com.arello-mobile:moxy:$moxyVersion"
@@ -50,5 +51,7 @@ dependencies {
     annotationProcessor "com.google.dagger:dagger-compiler:$daggerVersion"
 
     //Bottom Navigation Bar
-    compile 'com.ashokvarma.android:bottom-navigation-bar:1.3.0'
+    compile ('com.ashokvarma.android:bottom-navigation-bar:1.3.0') {
+        exclude group: "com.android.support", module: "design"
+    }
 }
